refactor: 마이페이지에 뷰티 프로필 요약 카드 추가 및 뷰티 프로필 조회 페이지 삭제
